+++ tests/test_storage.py
@@ -70,6 +70,7 @@ def test_get_path_from_infos():
 
 @mock.patch.dict(os.environ, {}, clear=True)
 def test_s3_missing_env():
+    disconnect_s3_clients()
     with pytest.raises(MissingEnvironmentError):
         get_data_str("s3://bucket/path/to/object")
 
@@ -809,7 +810,7 @@ def test_exists_s3_ok(mocked_s3_client):
         assert False, f"S3 exists raises an exception: {exc}"
 
     s3_instance.head_object.side_effect = botocore.exceptions.ClientError(
-        operation_name="InvalidKeyPair.Duplicate", error_response={"Error": {"Code": "NoSuchKey"}}
+        operation_name="InvalidKeyPair.Duplicate", error_response={"Error": {"Code": "404"}}
     )
     try:
         assert not exists("s3://bucket/object.ext")