+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.1.4
+
+### [Fixed]
+
+* Storage : la réponse à un HEAD (test existence en S3) donne un code 404 et non NoSuchKey (confusion avec la lecture d'objet)
+* RasterSet: le chargement d'un raster set à partir d'un fichier ou d'un descripteur utilise la librairie Storage et non la librairie GDAL
+
 ## 2.1.3
 
 ### [Fixed]